feat: delete page Suspense Bondary 래핑 (#41)
